Переносит тесты редьюсера
